Student population script done
Co-author: cherylwu834@gmail.com
